consolidate, organize org-side JS
