+++ templates/layouts/base.html
@@ -21,7 +21,7 @@
 
     <script src="/static/vendor/jquery-3.2.1/jquery-3.2.1.min.js"></script>
     <script src="/static/vendor/bootstrap-3.3.7/js/bootstrap.min.js"></script>
-    <script src="/static/js/base.js"></script>
+    <script src="/static/js/org.js"></script>
 
     {% block scriptContent %}
