+++ templates/organization/login.html
@@ -143,6 +143,3 @@
 
   </body>
   {% endblock %} 
-  {% block scriptContent %}
-    <script src="/static/js/auth.js"></script> 
-  {% endblock %}
